+++ changelog.md
@@ -2,3 +2,5 @@
 + Upgrade dependencies
 + Add `/music skip`
 + Add `/music loop`
++ Add support for `music.youtube.com` links
++ Add support for `youtu.be` links